mantenimiento cliente
eliminar , crear , modificar

+++ Proyecto Final/js/invoices.js
@@ -1,27 +1,25 @@
-var text = '{"clients":[]}';
-var Util = Util || {
-	agregarClient: function()
+var tex ;
+var Invoice = Invoice || {
+	agregarInvoice: function()
 	{	
-		var obj;
-		if (localStorage.getItem('Ul')=="") 
+		var o;
+		if (localStorage.getItem('inv')==null || localStorage.getItem('inv')=="") 
 		{
-			text='{"clients":[]}';
-			obj = JSON.parse(text);
+			tex='{"Invoices":[]}';
+			o = JSON.parse(tex);
 		}
 		else
 		{
-			obj=JSON.parse(localStorage.getItem('Ul'));
+			o=JSON.parse(localStorage.getItem('inv'));
 		}
-		//este mae convierte a string la variable global llamada text		
-		
 		//este mae agrega otro nuevo elemento		
-		obj ['clients'].push({"firstName":document.getElementById('first_name_client').value,"lastName":document.getElementById('last_name_client').value,"id":document.getElementById('id_client').value,"phone":document.getElementById('phone_client').value});
-		//lo convierte a objeto 
-		text = JSON.stringify (obj);
+		o ['Invoices'].push({"id":'1',"client":document.getElementById('client_invoice').value,"descript":document.getElementById('description_invoice').value,"mount":document.getElementById('mount_invoice').value});
+		//lo convierte a objeto text
+		tex = JSON.stringify (o);
 		//de nuevo a parse del nuevo elemento
-		obj = JSON.parse(text);
+		o = JSON.parse(tex);
 		//lo guarda en local storage
-		localStorage.setItem('Ul', JSON.stringify(obj));
+		localStorage.setItem('inv', JSON.stringify(o));
 	},
 	deleteClient: function()
 	{	
@@ -51,23 +49,33 @@ var Util = Util || {
 	};
 
 },
-llenartabla: function()
+llenarInvoice: function()
 {
-	var object2 = JSON.parse(localStorage.getItem('Ul'));
+	if (localStorage.getItem('inv')==null || localStorage.getItem('inv')=="") 
+	{
+		alert("No hay elementos en esta tabla para mostrar.");
+	}
+	else
+	{
+		var object2 = JSON.parse(localStorage.getItem('inv'));
+	//alert(object2.Invoices.length+"  cantida filas");
 		//ciclo para imprimir todo lo guardado
-		for (var i = 0; i < object2.clients.length; i++) {
-			var tbl = document.getElementById('tableclient');
+		for (var i = 0; i < object2.Invoices.length; i++) {
+			var tbl = document.getElementById('table_invoices');
 			var lastRow = tbl.rows.length;
 			var row = tbl.insertRow(lastRow);
 			var id = row.insertCell(0);
-			var fullname = row.insertCell(1);
-			var phone = row.insertCell(2);
-			var actions = row.insertCell(3);
-			id.innerHTML = object2.clients[i].id;
-			fullname.innerHTML = object2.clients[i].firstName+" "+object2.clients[i].lastName;
-			phone.innerHTML = object2.clients[i].phone;
+			var client = row.insertCell(1);
+			var description = row.insertCell(2);
+			var mount = row.insertCell(3);
+			var actions = row.insertCell(4);
+			id.innerHTML = object2.Invoices[i].id;
+			client.innerHTML = object2.Invoices[i].client;
+			description.innerHTML= object2.Invoices[i].descript;
+			mount.innerHTML= object2.Invoices[i].mount;
 			actions.innerHTML="<a  href='edit_client.html'><img  id='editar' src='edit.png'/> </a> <a  href='delete_client.html'><img  id='editar' src='delete.png'/> </a>";
 			};
-
 		}
-	};
+
+	}
+};
