+++ Proyecto Final/js/user.js
@@ -1,27 +1,24 @@
-var text = '{"clients":[]}';
-var Util = Util || {
-	agregarClient: function()
+var tex ;
+var Usuario = Usuario || {
+	agregarUser: function()
 	{	
-		var obj;
-		if (localStorage.getItem('Ul')=="") 
+		var ob;
+		if (localStorage.getItem('User')==null || localStorage.getItem('User')=="") 
 		{
-			text='{"clients":[]}';
-			obj = JSON.parse(text);
+			tex='{"User":[]}';
+			ob = JSON.parse(tex);
 		}
 		else
 		{
-			obj=JSON.parse(localStorage.getItem('Ul'));
+			ob=JSON.parse(localStorage.getItem('User'));
 		}
-		//este mae convierte a string la variable global llamada text		
-		
-		//este mae agrega otro nuevo elemento		
-		obj ['clients'].push({"firstName":document.getElementById('first_name_client').value,"lastName":document.getElementById('last_name_client').value,"id":document.getElementById('id_client').value,"phone":document.getElementById('phone_client').value});
-		//lo convierte a objeto 
-		text = JSON.stringify (obj);
+		ob ['User'].push({"fullname":document.getElementById('user_name').value,"username":document.getElementById('name_user').value,"pasword":document.getElementById('pasword_repeat').value});
+		//lo convierte a objeto text
+		tex = JSON.stringify (ob);
 		//de nuevo a parse del nuevo elemento
-		obj = JSON.parse(text);
+		ob = JSON.parse(tex);
 		//lo guarda en local storage
-		localStorage.setItem('Ul', JSON.stringify(obj));
+		localStorage.setItem('User', JSON.stringify(ob));
 	},
 	deleteClient: function()
 	{	
@@ -51,23 +48,32 @@ var Util = Util || {
 	};
 
 },
-llenartabla: function()
+llenarUser: function()
 {
-	var object2 = JSON.parse(localStorage.getItem('Ul'));
+	if (localStorage.getItem('User')==null || localStorage.getItem('User')=="") 
+	{
+		alert('No hay usuarios guardadas.');
+	}
+	else
+	{
+		var object2 = JSON.parse(localStorage.getItem('User'));
 		//ciclo para imprimir todo lo guardado
-		for (var i = 0; i < object2.clients.length; i++) {
-			var tbl = document.getElementById('tableclient');
+		for (var i = 0; i < object2.User.length; i++) {
+			var tbl = document.getElementById('table_user');
 			var lastRow = tbl.rows.length;
 			var row = tbl.insertRow(lastRow);
-			var id = row.insertCell(0);
-			var fullname = row.insertCell(1);
-			var phone = row.insertCell(2);
+			var full = row.insertCell(0);
+			var username = row.insertCell(1);
+			var pasword = row.insertCell(2);
 			var actions = row.insertCell(3);
-			id.innerHTML = object2.clients[i].id;
-			fullname.innerHTML = object2.clients[i].firstName+" "+object2.clients[i].lastName;
-			phone.innerHTML = object2.clients[i].phone;
+			full.innerHTML = object2.User[i].fullname;
+			username.innerHTML = object2.User[i].username;
+			pasword.innerHTML= object2.User[i].pasword;
 			actions.innerHTML="<a  href='edit_client.html'><img  id='editar' src='edit.png'/> </a> <a  href='delete_client.html'><img  id='editar' src='delete.png'/> </a>";
+			//actions.innerHTML="<img src='edit.png' href='edit_client.html'/> <img src='delete.png' href='delete_client.html'/>";
 			};
 
+
 		}
-	};
+	}
+};
